+++ lib/math/bitset_square_matrix.hpp
@@ -7,25 +7,25 @@ private:
   vector<Bitset> val;
 
 public:
-  BitsetSquareMatrix(size_t Nb) : val(Nb, Bitset(Nb)) {}
+  BitsetSquareMatrix(int Nb) : val(Nb, Bitset(Nb)) {}
 
   Bitset& operator[](int n) {
     return val[n];
   }
 	
 	BitsetSquareMatrix operator+=(const BitsetSquareMatrix& m) {
-    for (size_t i = 0; i < size(); ++i) val[i] ^= const_cast<BitsetSquareMatrix&>(m)[i];
+    for (int i = 0; i < size(); ++i) val[i] ^= const_cast<BitsetSquareMatrix&>(m)[i];
 		return *this;
 	}
 	
 	BitsetSquareMatrix operator-=(const BitsetSquareMatrix& m) {
-    for (size_t i = 0; i < size(); ++i) val[i] ^= const_cast<BitsetSquareMatrix&>(m)[i];
+    for (int i = 0; i < size(); ++i) val[i] ^= const_cast<BitsetSquareMatrix&>(m)[i];
 		return *this;
 	}
 
   BitsetSquareMatrix operator*=(const BitsetSquareMatrix& _m) {
     BitsetSquareMatrix m = _m.transpose(), res(size());
-    for (size_t i = 0; i < size(); ++i) res[i] = *this * m[i];
+    for (int i = 0; i < size(); ++i) res[i] = *this * m[i];
     return *this = res.transpose();
   }
 
@@ -40,42 +40,42 @@ public:
 
   Bitset operator*(const Bitset& v) const {
     Bitset res(size());
-    for (size_t i = 0; i < size(); ++i) res[i] = (val[i] & v).parity();
+    for (int i = 0; i < size(); ++i) res[i] = (val[i] & v).parity();
     return res;
   }
 
   BitsetSquareMatrix transpose() const {
     BitsetSquareMatrix res(size());
-    for (size_t i = 0; i < size(); ++i) {
-      for (size_t j = 0; j < size(); ++j) {
+    for (int i = 0; i < size(); ++i) {
+      for (int j = 0; j < size(); ++j) {
         res[i][j] = const_cast<BitsetSquareMatrix&>(*this)[j][i];
       }
     }
     return res;
   }
 
-  size_t size() const {
+  int size() const {
     return val.size();
   }
 
   BitsetSquareMatrix identity() const {
     BitsetSquareMatrix res(size());
-    for (size_t i = 0; i < size(); ++i) res[i][i] = true;
+    for (int i = 0; i < size(); ++i) res[i][i] = true;
     return res;
   }
 
   BitsetSquareMatrix inverse() const {
-    size_t n = size();
+    int n = size();
     BitsetSquareMatrix mat = *this;
     BitsetSquareMatrix inv = identity();
-    for (size_t i = 0; i < n; ++i) {
+    for (int i = 0; i < n; ++i) {
       int p = i;
-      for (size_t j = i + 1; j < n; ++j) {
+      for (int j = i + 1; j < n; ++j) {
         if (abs(mat[j][i]) > abs(mat[p][i])) p = j;
       }
       swap(mat[i], mat[p]);
       swap(inv[i], inv[p]);
-      for (size_t j = 0; j < n; ++j) {
+      for (int j = 0; j < n; ++j) {
         if (i == j) continue;
         if (!mat[j][i]) continue;
         mat[j] ^= mat[i];
