+++ lib/math/permutation_matrix.hpp
@@ -6,7 +6,7 @@ private:
   vector<int> val;
 
 public:
-  PermutationMatrix(size_t Nb) : val(Nb) {}
+  PermutationMatrix(int Nb) : val(Nb) {}
 
   int& operator[](int n) {
     return val[n];
@@ -14,7 +14,7 @@ public:
 	
   PermutationMatrix operator*=(const PermutationMatrix& m) {
     PermutationMatrix res(size());
-    for (size_t i = 0; i < size(); ++i) res[i] = val[const_cast<PermutationMatrix&>(m)[i]];
+    for (int i = 0; i < size(); ++i) res[i] = val[const_cast<PermutationMatrix&>(m)[i]];
     return *this = res;
   }
 
@@ -22,19 +22,19 @@ public:
 		return *this *= m.inverse();
 	}
 
-  size_t size() const {
+  int size() const {
     return val.size();
   }
 
   PermutationMatrix identity() const {
     PermutationMatrix res(size());
-    for (size_t i = 0; i < size(); ++i) res[i] = i;
+    for (int i = 0; i < size(); ++i) res[i] = i;
     return res;
   }
 
   PermutationMatrix inverse() const {
     PermutationMatrix inv(size());
-    for (size_t i = 0; i < size(); ++i) inv[val[i]] = i;
+    for (int i = 0; i < size(); ++i) inv[val[i]] = i;
     return inv;
   }
 };
